Memory Beta site search

Perform a Memory Beta site search.

mb_search(text, browse = FALSE)

Arguments

text

character, search query.

browse

logical, open search results page in browser.

Value

a data frame

Details

This function returns a data frame containing the title, truncated text preview, and relative URL for the first page of search results. It does not recursively collate search results through subsequent pages of results. There could be an unexpectedly high number of pages of results depending on the search query. Since the general nature of this search feature seems relatively casual anyway, it aims only to provide a first page preview.

Examples

mb_search("Worf")
#> # A tibble: 25 × 3
#>    title                  text                                             url  
#>    <chr>                  <chr>                                            <chr>
#>  1 Worf                   Worf, son of Mogh, of the House of Martok, (bor… http…
#>  2 Worf, father of Mogh   Worf was a Klingon male who lived during the la… http…
#>  3 Worf (1ST)             In the First Splinter timeline, Worf, son of Mo… http…
#>  4 Worf (mirror)          In the mirror universe, Worf, son of Mogh was t… http…
#>  5 Worf's First Adventure Before he was an officer aboard the U.S.S. Ente… http…
#>  6 Worf's family          of Martok. In 2386, Worf's marriage to Grilka b… http…
#>  7 Worf (alternates)      There have been multiple timelines, alternate r… http…
#>  8 Worf Special           The Worf Special was an issue of Malibu Comics'… http…
#>  9 Worf's bird-of-prey    Eridani system and began an occupation of plane… http…
#> 10 Personal log, Worf     The Personal log of Worf contained several log … http…
#> # ℹ 15 more rows