Go to Wikipedia entry for a specific book series

This function opens a browser tab to the main Wikipedia entry for all Star Trek novels automatically scrolled to the section pertaining to id. To see the available IDs, call st_book_series with no arguments.

st_book_series(id)

Arguments

id
character, abbreviation for a series. See details.

Value

opens a browser tab, nothing is returned unless id is not provided, in which case a data frame is returned.

See also

Examples

st_book_series()

#> # A tibble: 17 x 2
#>    series                                abb       
#>    <chr>                                 <chr>     
#>  1 THe Original Series                   TOS       
#>  2 The Next Generation                   TNG       
#>  3 Deep Space Nine                       DS9       
#>  4 Voyager                               VOY       
#>  5 Enterprise                            ENT       
#>  6 Discovery                             DSC       
#>  7 New Frontier                          NF        
#>  8 Stargazer                             SG        
#>  9 IKS Gorkon/Klingon Empire             IKE       
#> 10 Titan                                 TIT       
#> 11 Vanguard                              VAN       
#> 12 Seekers                               SKR       
#> 13 Mini-series                           miniseries
#> 14 Starfleet Corps of Engineers          SCE       
#> 15 Department of Temporal Investigations DTI       
#> 16 Mirror Universe                       MIR       
#> 17 Starfleet Academy                     SFA       

st_book_series("DS9")