Ancillary location data for map tiles

Obtain a table of ancillary data associated with various locations of interest, given a specific map tile set ID.

st_tiles_data(id)

Arguments

id
character, name of a map tile set.

Value

a data frame

Details

This function returns a small example data frame of location-specific data along with grid cell coordinates that are specific to the requested map tile set ID.

See also

Examples

st_tiles_data("galaxy2")

#> # A tibble: 9 x 8
#>   id      loc         col   row body   category zone                   species  
#>   <chr>   <chr>     <dbl> <dbl> <chr>  <chr>    <chr>                  <chr>    
#> 1 galaxy2 Earth      2201  1595 Planet Homewor~ United Federation of ~ Human    
#> 2 galaxy2 Romulus    2514  1178 Planet Homewor~ Romulan Star Empire    Romulan  
#> 3 galaxy2 Qo'noS     3197  2303 Planet Homewor~ Klingon Empire         Klingon  
#> 4 galaxy2 Breen      1228  1181 Planet Homewor~ Breen Confederacy      Breen    
#> 5 galaxy2 Ferengin~  2026   886 Planet Homewor~ Ferengi Alliance       Ferengi  
#> 6 galaxy2 Cardassia  1543  1903 Planet Homewor~ Cardassian Union       Cardassi~
#> 7 galaxy2 Tholia      713  2971 Planet Homewor~ Tholian Assembly       Tholian  
#> 8 galaxy2 Tzenketh   1734  1721 Planet Homewor~ Tzenkethi Coalition    Tzenkethi
#> 9 galaxy2 Talar      1338  2368 Planet Homewor~ Talarian Republic      Talarian