Memory Beta site search

Perform a Memory Beta site search.

mb_search(text, browse = FALSE)

Arguments

text
character, search query.

browse
logical, open search results page in browser.

Value

a data frame

Details

This function returns a data frame containing the title, truncated text preview, and relative URL for the first page of search results. It does not recursively collate search results through subsequent pages of results. There could be an unexpectedly high number of pages of results depending on the search query. Since the general nature of this search feature seems relatively casual anyway, it aims only to provide a first page preview.

Examples

if(has_internet()) mb_search("Worf")

#> # A tibble: 25 x 3
#>    title                text                                 url                
#>    <chr>                <chr>                                <chr>              
#>  1 Worf, son of Mogh    https://memory-beta.fandom.com/wiki~ Worf,_son_of_Mogh  
#>  2 Worf (mirror)        https://memory-beta.fandom.com/wiki~ Worf_(mirror)      
#>  3 Worf's family        https://memory-beta.fandom.com/wiki~ Worf%27s_family    
#>  4 Worf, father of Mogh https://memory-beta.fandom.com/wiki~ Worf,_father_of_Mo~
#>  5 Worf's First Advent~ https://memory-beta.fandom.com/wiki~ Worf%27s_First_Adv~
#>  6 Worf Special         https://memory-beta.fandom.com/wiki~ Worf_Special       
#>  7 Worf's bird-of-prey  https://memory-beta.fandom.com/wiki~ Worf%27s_bird-of-p~
#>  8 Mogh, son of Worf    https://memory-beta.fandom.com/wiki~ Mogh,_son_of_Worf  
#>  9 K'Dhan, son of Worf  https://memory-beta.fandom.com/wiki~ K%27Dhan,_son_of_W~
#> 10 Category:Memory Bet~ https://memory-beta.fandom.com/wiki~ Category:Memory_Be~
#> # ... with 15 more rows