Retrieve Star Trek data from STAPI

Retrieve Star Trek data from the Star Trek API (STAPI).

stapi(id, page = 1, uid = NULL, page_count = FALSE)

Arguments

id
character, name of STAPI entity. See details.

page
integer vector, defaults to first page.

uid
NULL for search mode, character for extraction mode. See details.

page_count
logical, set to TRUE to do a preliminary check of the total number a pages of results available for a potential entity search. This will only have the impact of searching the first page.

Value

a data frame in search mode, a list in extraction mode, and nothing is returned in page count check mode but the result is printed to the console.

Details

See stapiEntities for all the currently available API entities. These are the IDs for dataset collections or categories passed to id.

The universal ID uid can be supplied to retrieve a more specific subset of data. By default, uid = NULL and stapi operates in search mode. As part of a stepwise process, you can first use search mode. Then if the resulting data frame includes a uid column, you can make a second call to the function providing a specific uid. This puts stapi into extraction mode and will return satellite data associated with the unique entry from the original general sweep of the entity id.

rtrek employs anti-DDOS measures. It will not perform an API call to STAPI more than once per second. To be an even better neighbor, you can increase this wait time using options, e.g. options(rtrek_antiddos = 10) to increase the minimum time between API calls to ten seconds. Values less than one are ignored (defaulting back to one second) and a warning will be thrown when making any API call if this is the case.

Currently STAPI contains primarily real world data such as episode air dates, movie metadata, or production company information. Fictional world data is secondary and more limited.

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

stapi("character", page_count = TRUE) # check first

#> Total pages to retrieve all results: 64 

stapi("character", page = 2) %>% select(1:2)

#> # A tibble: 100 x 2
#>    uid            name         
#>    <chr>          <chr>        
#>  1 CHMA0000042544 Stevie Wonder
#>  2 CHMA0000018397 Yeggie       
#>  3 CHMA0000007954 Arex         
#>  4 CHMA0000026438 Joseph       
#>  5 CHMA0000232412 J. Zimmerman 
#>  6 CHMA0000014055 Doyle        
#>  7 CHMA0000012951 Butler       
#>  8 CHMA0000048382 Lito         
#>  9 CHMA0000190073 B. Mooneyham 
#> 10 CHMA0000171058 Anna Dixon   
#> # ... with 90 more rows

Q <- stapi("character", uid = "CHMA0000025118")
Q$episodes %>% select(uid, title, stardateFrom, stardateTo)

#>              uid                 title stardateFrom stardateTo
#> 1 EPMA0000001458    All Good Things...      47988.0    47988.0
#> 2 EPMA0000001329                 Q Who      42761.3    42761.3
#> 3 EPMA0000001377                  Qpid      44741.9    44741.9
#> 4 EPMA0000000483 Encounter at Farpoint      41153.7    41153.7
#> 5 EPMA0000000651              Tapestry           NA         NA
#> 6 EPMA0000000845                Q-Less      46531.2    46531.2
#> 7 EPMA0000162588            Death Wish           NA         NA
#> 8 EPMA0000001413                True Q      46192.3    46192.3
#> 9 EPMA0000001510    The Q and the Grey      50384.2    50392.7