Retrieve Star Trek data from STAPI

Retrieve Star Trek data from the Star Trek API (STAPI).

stapi(id, page = 1, uid = NULL, page_count = FALSE)

Arguments

id
character, name of STAPI entity. See details.

page
integer vector, defaults to first page.

uid
NULL for search mode, character for extraction mode. See details.

page_count
logical, set to TRUE to do a preliminary check of the total number a pages of results available for a potential entity search. This will only have the impact of searching the first page.

Value

a data frame in search mode, a list in extraction mode, and nothing is returned in page count check mode but the result is printed to the console.

Details

See stapiEntities for all the currently available API entities. These are the IDs for dataset collections or categories passed to id.

The universal ID uid can be supplied to retrieve a more specific subset of data. By default, uid = NULL and stapi operates in search mode. As part of a stepwise process, you can first use search mode. Then if the resulting data frame includes a uid column, you can make a second call to the function providing a specific uid. This puts stapi into extraction mode and will return satellite data associated with the unique entry from the original general sweep of the entity id.

rtrek employs anti-DDOS measures. It will not perform an API call to STAPI more than once per second. To be an even better neighbor, you can increase this wait time using options, e.g. options(rtrek_antiddos = 10) to increase the minimum time between API calls to ten seconds. Values less than one are ignored (defaulting back to one second) and a warning will be thrown when making any API call if this is the case.

Examples

library(dplyr)
stapi("character", page_count = TRUE) # check first

#> Total pages to retrieve all results: 62 

stapi("character", page = 2) %>% select(1:2)

#> # A tibble: 100 x 2
#>    uid            name                  
#>    <chr>          <chr>                 
#>  1 CHMA0000016634 Fuller                
#>  2 CHMA0000022998 Burkus                
#>  3 CHMA0000214779 Masaka's city         
#>  4 CHMA0000040790 Thorne                
#>  5 CHMA0000024856 Ah-Kel                
#>  6 CHMA0000171279 Robert Justman Raymond
#>  7 CHMA0000025118 Q                     
#>  8 CHMA0000069349 John Dwyer            
#>  9 CHMA0000057119 Louis of France       
#> 10 CHMA0000021854 Marat Kobar           
#> # ... with 90 more rows

Q <- stapi("character", uid = "CHMA0000025118")
Q$episodes %>% select(uid, title, stardateFrom, stardateTo)

#>              uid                 title stardateFrom stardateTo
#> 1 EPMA0000001458    All Good Things...      47988.0    47988.0
#> 2 EPMA0000001329                 Q Who      42761.3    42761.3
#> 3 EPMA0000000483 Encounter at Farpoint      41153.7    41153.7
#> 4 EPMA0000162588            Death Wish           NA         NA
#> 5 EPMA0000001510    The Q and the Grey      50384.2    50392.7
#> 6 EPMA0000000651              Tapestry           NA         NA
#> 7 EPMA0000000845                Q-Less      46531.2    46531.2
#> 8 EPMA0000001413                True Q      46192.3    46192.3
#> 9 EPMA0000001377                  Qpid      44741.9    44741.9